Source: https://github.com/msupply-foundation/open-msupply/commit/6cb95690066d8503c9512128df9209b84addc61f

+++ server/repository/src/db_diesel/stocktake_line_row.rs
@@ -17,8 +17,8 @@ table! {
         stock_line_id -> Nullable<Text>,
         location_id	-> Nullable<Text>,
         comment	-> Nullable<Text>,
-        snapshot_number_of_packs -> Integer,
-        counted_number_of_packs -> Nullable<Integer>,
+        snapshot_number_of_packs -> Double,
+        counted_number_of_packs -> Nullable<Double>,
 
         // stock line related fields:
         item_id -> Text,
@@ -36,6 +36,7 @@ joinable!(stocktake_line -> stocktake (stocktake_id));
 joinable!(stocktake_line -> stock_line (stock_line_id));
 
 #[derive(Clone, Queryable, Insertable, AsChangeset, Debug, PartialEq, Default)]
+#[changeset_options(treat_none_as_null = "true")]
 #[table_name = "stocktake_line"]
 pub struct StocktakeLineRow {
     pub id: String,
@@ -45,8 +46,8 @@ pub struct StocktakeLineRow {
     pub location_id: Option<String>,
     /// Comment for this stocktake line
     pub comment: Option<String>,
-    pub snapshot_number_of_packs: i32,
-    pub counted_number_of_packs: Option<i32>,
+    pub snapshot_number_of_packs: f64,
+    pub counted_number_of_packs: Option<f64>,
 
     // stock line related fields:
     /// When a creating a new stock line this field holds the required item id
